services/booking-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic-settings==2.1.0
python-json-logger==2.0.7
prometheus-fastapi-instrumentator==7.0.0
google-cloud-pubsub==2.19.0

